+++ foreign/go/contracts/topics.go
@@ -21,7 +21,7 @@ import "time"
 
 type CreateTopicRequest struct {
 	StreamId             Identifier    `json:"streamId"`
-	TopicId              int           `json:"topicId"`
+	TopicId              uint32        `json:"topicId"`
 	PartitionsCount      int           `json:"partitionsCount"`
 	CompressionAlgorithm uint8         `json:"compressionAlgorithm"`
 	MessageExpiry        time.Duration `json:"messageExpiry"`
@@ -41,8 +41,8 @@ type UpdateTopicRequest struct {
 }
 
 type Topic struct {
-	Id                   int           `json:"id"`
-	CreatedAt            int           `json:"createdAt"`
+	Id                   uint32        `json:"id"`
+	CreatedAt            uint64        `json:"createdAt"`
 	Name                 string        `json:"name"`
 	SizeBytes            uint64        `json:"sizeBytes"`
 	MessageExpiry        time.Duration `json:"messageExpiry"`
@@ -50,7 +50,7 @@ type Topic struct {
 	MaxTopicSize         uint64        `json:"maxTopicSize"`
 	ReplicationFactor    uint8         `json:"replicationFactor"`
 	MessagesCount        uint64        `json:"messagesCount"`
-	PartitionsCount      int           `json:"partitionsCount"`
+	PartitionsCount      uint32        `json:"partitionsCount"`
 }
 
 type TopicDetails struct {
